refactor: remove TUI module, make headless the default mode
Remove the ratatui/crossterm terminal UI entirely. The project now has
two UI modes: headless REPL (default via `cargo run`) and Tauri desktop
GUI (`cargo tauri dev`).

- Extract App and ScrollState from src/tui/ into src/app.rs
- Delete src/tui/ (mod.rs, debug_panel.rs) and docs/design/tui-design.md
- Remove ratatui + crossterm dependencies from Cargo.toml
- Refactor src/loading.rs to use (u8,u8,u8) tuples instead of ratatui Color
- Remove --tui CLI flag; headless runs by default without flags
- Remove dead code from main.rs (TUI game loop, handle_system_command, etc.)
- Update all doc comments and documentation to reflect removal
