+++ Cargo.toml
@@ -13,8 +13,6 @@ default-run = "parish"
 [dependencies]
 parish-core = { path = "crates/parish-core" }
 tokio = { version = "1", features = ["full"] }
-ratatui = "0.29"
-crossterm = "0.28"
 clap = { version = "4", features = ["derive", "env"] }
 anyhow = "1"
 thiserror = "2"
